Collapse network factory hierarchy
